vid.width -> vid.displaywidth, vid.height -> vid.displayheight
The vid struct has both width and height variables but also conwidth and
conheight variables. These names are however very misleading as the conwidth
and conheight variables essentially weren't used and width and height meant
conwidth and conheight.

This change changes the name of the width and height to displaywidth and
displayheight, making them mean the pixel size of the display, and makes
conwidth and conheight mean the pixel size of the console and other 2D
elements drawn.